Corrected syntax errors in search procedure.
I corrected a few syntax errors found throughout the search procedure,
as well as in related procedures. The problem was that I was using my
own "get" procedure incorrectly and forgetting that "get" returns a
procedure that, given a list, returns the member of that list at the
given row and column.