Add path retracing to search procedure.
The search procedure will now retrace the path from goal to start.
However, there are still errors with the syntax -- specifically, at line
207, it says that '() is not an object. This is because by default, the
parent member of the tile class is set to '(). However, this is supposed
to be modified when the algorithm is run. But since the algorithm isn't
even being executed yet, this means Racket is assuming the value of
parent before the code is even executed.

I'm not sure how I should assign a default value while at the same time
being able to later assign a reference to the parent tile so that I can
properly send method requests to the parent of each tile.